Stop t-SNE from overflowing the stack.
This happens because the TSNE workers is throwing out embeddingUpdate signals in a parallel thread.
The main thread then calls processEvents on a regular basis, which handles the event, but this handler calls processEvents again.
As there are plenty of signals to process (coming from the parallel worker) it will just keep recursing.